# Improve speed of `group_by()`