--- a/R/group_utils.R
+++ b/R/group_utils.R
@@ -63,16 +63,18 @@ calculate_groups <- function(data, groups, drop = group_by_drop_default(data)) {
   unique_groups <- unique(data[, groups, drop = FALSE])
   is_factor <- do.call(c, lapply(unique_groups, function(x) is.factor(x)))
   n_comb <- nrow(unique_groups)
-  rows <- rep(list(NA), n_comb)
-  data_groups <- interaction(data[, groups, drop = TRUE])
   
+  temp_id <- paste(sample(letters), collapse = "")
+
   # Concatenate group variables
-  pasted_groups <- do.call(paste, c(unique_groups[, groups, drop = FALSE], sep = "."))
-  pasted_groups[is.na(unique_groups)] <- NA
-  
+  pasted_groups <- do.call(paste, c(unique_groups, sep = "."))
+  data[[temp_id]] <- do.call(paste, c(data[, groups, drop = FALSE], sep = "."))
+
+  rows <- rep(list(NA), n_comb)
   for (i in seq_len(n_comb)) {
-    rows[[i]] <- which(data_groups %in% pasted_groups[i])
+    rows[[i]] <- which(data[[temp_id]] %in% pasted_groups[i])
   }
+  data[[temp_id]] <- NULL
 
   if (!isTRUE(drop) && any(is_factor)) {
     na_lvls <- do.call(