@etiennebacher 2023-01-06
TODO: the following code doesn't produce the right output (note that the output is also wrong without this PR):
``` r
suppressPackageStartupMessages(library(poorman))

d <- data.frame(
  orig = rep(c("France", "UK"), each = 4),
  dest = rep(c("Spain", "Germany"), times = 4),
  year = rep(rep(c(2010, 2011), each = 2), 2),
  value = 1:8
)
d[2, 1] <- NA
d[7, 2] <- NA

d
#>     orig    dest year value
#> 1 France   Spain 2010     1
#> 2   <NA> Germany 2010     2
#> 3 France   Spain 2011     3
#> 4 France Germany 2011     4
#> 5     UK   Spain 2010     5
#> 6     UK Germany 2010     6
#> 7     UK    <NA> 2011     7
#> 8     UK Germany 2011     8

d %>%
  group_by(orig, dest) %>%
  group_data()
#>     orig    dest .rows
#> 1 France Germany     4
#> 2 France   Spain  1, 3
#> 3     UK Germany  6, 8
#> 4     UK   Spain     5
#> 5     UK    <NA>      
#> 6   <NA> Germany  2, 7
```

The groups UK-NA and NA-Germany should have 1 row each.